telegram_forward_bot/requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
aiosqlite==0.19.0
APScheduler==3.10.4
googletrans==4.0.0rc1
Pillow==10.1.0
python-dotenv==1.0.0
pytz==2023.3
aiofiles==23.2.1
